fix unsigned vs signed compare

+++ src/Adafruit_MCPSRAM.cpp
@@ -224,7 +224,7 @@ void Adafruit_MCPSRAM::read(uint16_t addr, uint8_t *buf, uint16_t num,
   }
 
   // read data into buffer
-  for (int i = 0; i < num; i++) {
+  for (uint16_t i = 0; i < num; i++) {
 
     if (hwSPI) {
       buf[i] = _spi->transfer(0x00);
